requirements.txt
Flask==2.0.2
gunicorn==20.1.0
cachetools==5.2.1    
certifi==2022.12.7
charset-normalizer==3.0.1    
contourpy==1.0.7    
cycler==0.11.0   
fonttools==4.38.0   
google-api-core==2.11.0   
google-api-python-client==2.72.0   
google-auth==2.16.0   
google-auth-httplib2==0.1.0    
google-cloud==0.34.0   
google-cloud-appengine-logging==1.3.0    
google-cloud-audit-log==0.2.5    
google-cloud-core==2.3.2    
google-cloud-logging==3.4.0    
googleapis-common-protos==1.58.0   
grpc-google-iam-v1==0.12.6   
grpcio==1.51.1   
grpcio-status==1.51.1   
httplib2==0.21.0   
idna==3.4      
kiwisolver==1.4.4    
matplotlib==3.6.3    
numpy==1.24.1   
packaging==23.0     
pandas==1.5.2    
Pillow==9.4.0    
pip ==19.2.3   
pretty-html-table==0.9.16   
proto-plus==1.22.2   
protobuf==4.21.12  
pyasn1==0.4.8    
pyasn1-modules==0.2.8    
pyparsing==3.0.9    
python-dateutil==2.8.2    
pytz==2022.7.1 
requests==2.28.2   
rsa==4.9      
seaborn==0.12.2   
setuptools==41.2.0   
six==1.16.0   
uritemplate==4.1.1    
urllib3==1.26.14  